Consolidate named-arg helpers into util
Hoist the byte-identical getNamespaceCallableBrsName and
getClassConstructorParams helpers (previously duplicated across
ScopeValidator, BrsFilePreTranspileProcessor, and CodeActionsProcessor)
into util as resolveNamespaceCallableName and getClassConstructorParams.
