+++ src/bscPlugin/transpile/BrsFilePreTranspileProcessor.ts
@@ -9,7 +9,7 @@ import type { Expression, Statement } from '../../parser/AstNode';
 import type { CallExpression, FunctionExpression, FunctionParameterExpression, NamedArgumentExpression, TernaryExpression } from '../../parser/Expression';
 import { LiteralExpression } from '../../parser/Expression';
 import { ParseMode } from '../../parser/Parser';
-import type { ClassStatement, ConstStatement, IfStatement, MethodStatement, NamespaceStatement } from '../../parser/Statement';
+import type { ConstStatement, IfStatement, NamespaceStatement } from '../../parser/Statement';
 import type { Scope } from '../../Scope';
 import util from '../../util';
 
@@ -50,7 +50,7 @@ export class BrsFilePreTranspileProcessor {
                     params = callableContainerMap.get(funcName.toLowerCase())?.[0]?.callable?.functionStatement?.func?.parameters;
                 } else if (isDottedGetExpression(callExpr.callee)) {
                     // Namespace function call (e.g. MyNs.myFunc(a: 1))
-                    const brsName = this.getNamespaceCallableBrsName(callExpr.callee, scope);
+                    const brsName = util.resolveNamespaceCallableName(callExpr.callee, scope);
                     if (brsName) {
                         params = callableContainerMap.get(brsName.toLowerCase())?.[0]?.callable?.functionStatement?.func?.parameters;
                     }
@@ -81,7 +81,7 @@ export class BrsFilePreTranspileProcessor {
                     continue;
                 }
 
-                const params = this.getClassConstructorParams(classLink.item, scope, containingNamespace);
+                const params = util.getClassConstructorParams(classLink.item, scope, containingNamespace);
                 this.rewriteNamedArgCall(callExpr, params);
             }
         }
@@ -121,40 +121,6 @@ export class BrsFilePreTranspileProcessor {
         }
 
         this.event.editor.arraySplice(callExpr.args, 0, callExpr.args.length, ...finalArgs);
-    }
-
-    /**
-     * If the callee is a dotted-get expression whose leftmost identifier is a known namespace,
-     * returns the BrightScript-flattened name (e.g. "MyNs_myFunc"). Otherwise returns undefined.
-     */
-    private getNamespaceCallableBrsName(callee: Expression, scope: Scope): string | undefined {
-        const parts = util.getAllDottedGetParts(callee);
-        if (!parts || !scope.namespaceLookup.has(parts[0].text.toLowerCase())) {
-            return undefined;
-        }
-        return parts.map(p => p.text).join('_');
-    }
-
-    /**
-     * Walk the class and its ancestor chain to find the first constructor's parameter list.
-     * Returns an empty array if no constructor is defined anywhere in the hierarchy.
-     */
-    private getClassConstructorParams(classStmt: ClassStatement, scope: Scope, containingNamespace: string | undefined): FunctionParameterExpression[] {
-        let stmt: ClassStatement | undefined = classStmt;
-        while (stmt) {
-            const ctor = stmt.body.find(
-                s => (s as MethodStatement)?.name?.text?.toLowerCase() === 'new'
-            ) as MethodStatement | undefined;
-            if (ctor) {
-                return ctor.func.parameters;
-            }
-            if (!stmt.parentClassName) {
-                break;
-            }
-            const parentName = stmt.parentClassName.getName(ParseMode.BrighterScript);
-            stmt = scope.getClassFileLink(parentName, containingNamespace)?.item;
-        }
-        return [];
     }
 
     private iterateExpressions() {
